Version 1.1
Fixed a couple of small issues:
Files with spaces in their names / paths are now quoted before being given as arguments to ctags
(though I plan to change this code from using os.popen to using os.execl so as to completely avoid these kinds of issues)
If the tags file is found in \ (e.g. c:\) then I was making the relative path incorrectly. This is fixed.
I've cleaned up the handling of '/' and '\&#039 in paths (since tags files specify files with '/' but your OS may not)